Per issue #21, better chcek the JWT claims to make sure that the user has access to what they're requesting

+++ api/ammo.go
@@ -6,9 +6,9 @@ import (
 
 	"strconv"
 	"time"
-
 	"atfgundb.com/app/db"
 	"atfgundb.com/app/models"
+	"atfgundb.com/app/routing"
 	"github.com/gin-gonic/gin"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 )
@@ -16,9 +16,12 @@ import (
 func RemoveAmmo(c *gin.Context) {
 	ammoId := c.Query("ammo_id")
 
-	db.RemoveAmmo(ammoId)
-
-	c.JSON(http.StatusOK, "{success: true}")
+	if (db.UserOwnsAmmo(ammoId, routing.UserId)) {
+		db.RemoveAmmo(ammoId)
+		c.JSON(http.StatusOK, "{success: true}")
+	} else {
+		c.JSON(http.StatusUnauthorized, "{error: 'Unauthorized'}")
+	}
 }
 
 func AddAmmoPurchase(c *gin.Context) {
@@ -28,29 +31,37 @@ func AddAmmoPurchase(c *gin.Context) {
 		log.Fatal("Unable to BindJSON")
 	}
 
-	ammoPurchase.DatePurchased = primitive.NewDateTimeFromTime(time.Now())
-	db.InsertAmmoPurchase(&ammoPurchase)
+	if (db.UserOwnsAmmo(ammoPurchase.AmmoId, routing.UserId)) {
+		ammoPurchase.DatePurchased = primitive.NewDateTimeFromTime(time.Now())
+		db.InsertAmmoPurchase(&ammoPurchase)
 
-	c.JSON(http.StatusOK, "{success: true}")
+		c.JSON(http.StatusOK, "{success: true}")
+	} else {
+		c.JSON(http.StatusUnauthorized, "{error: 'Unauthorized'}")
+	}
+	
 }
 
 func DisposeAmmo(c *gin.Context) {
 	ammoId := c.Query("ammo_id")
 	quantity := c.Query("quantity")
 
-	// convert quantity to int
-	squantity, _ := strconv.Atoi(quantity)
+	if (db.UserOwnsAmmo(ammoId, routing.UserId)) {
+		// convert quantity to int
+		squantity, _ := strconv.Atoi(quantity)
 
-	consumeAmmo(ammoId, squantity)
+		consumeAmmo(ammoId, squantity)
 
-	c.JSON(http.StatusOK, "{success: true}")
+		c.JSON(http.StatusOK, "{success: true}")
+	} else {
+		c.JSON(http.StatusUnauthorized, "{error: 'Unauthorized'}")
+	}
 }
 
 // would have loved to done this as a method on AmmoPurchase, but that was causing circular import issues
 // since db and model are separate packages
 func consumeAmmo(ammoId string, quantity int) error {
 	ammoPurchases := db.GetAmmoActivePurchases(ammoId)
-
 	quantityConsumed := int(0)
 
 	for _, ammoPurchase := range ammoPurchases {
@@ -89,9 +100,13 @@ func AddAmmo(c *gin.Context) {
 		log.Fatal("Unable to BindJSON")
 	}
 
-	log.Println("Calling insertupdategun")
-	db.InsertUpdateAmmo(&ammo)
-	c.JSON(http.StatusOK, "{success: true}")
+	if ammo.UserID != routing.UserId {
+		c.JSON(http.StatusUnauthorized, "{error: 'Unauthorized'}")
+	} else {
+		db.InsertUpdateAmmo(&ammo)
+		c.JSON(http.StatusOK, "{success: true}")
+	}
+	
 }
 
 func ListAmmo(c *gin.Context) {
